+++ pyembed/src/test/test_importer_module_loading.py
@@ -58,12 +58,45 @@ class TestImporterModuleLoading(unittest.TestCase):
         self.assertIsNone(f.find_spec("my_package", None))
 
     def test_source_package(self):
-        self._make_package("my_package")
+        p = self._make_package("my_package")
+
+        with (p / "__init__.py").open("wb") as fh:
+            fh.write(b"import io\n")
 
         f = self._finder_from_td()
 
         spec = f.find_spec("my_package", None)
-        self.assertIsNone(spec)
+        self.assertIsInstance(spec, importlib.machinery.ModuleSpec)
+        self.assertEqual(spec.name, "my_package")
+        self.assertIsInstance(spec.loader, OxidizedFinder)
+        self.assertIsNone(spec.origin)
+        self.assertIsNone(spec.loader_state)
+        self.assertIsInstance(spec.submodule_search_locations, list)
+        self.assertEqual(
+            spec.submodule_search_locations, [os.path.join(sys.argv[0], "my_package")],
+        )
+
+        # Default module creation semantics for source modules.
+        self.assertIsNone(f.create_module(spec))
+
+        m = importlib.util.module_from_spec(spec)
+        self.assertEqual(m.__name__, "my_package")
+        self.assertIsInstance(m.__loader__, OxidizedFinder)
+        self.assertEqual(m.__loader__, f)
+        self.assertEqual(m.__package__, "my_package")
+        self.assertEqual(m.__path__, [os.path.join(sys.argv[0], "my_package")])
+        self.assertFalse(hasattr(m, "__file__"))
+        self.assertFalse(hasattr(m, "__cached__"))
+
+        self.assertIsNone(f.exec_module(m))
+
+        self.assertEqual(f.get_source("my_package"), "import io\n")
+
+        code = compile(f.get_source("my_package"), "my_package", "exec")
+        self.assertEqual(f.get_code("my_package"), code)
+
+        with self.assertRaises(ImportError):
+            f.get_filename("my_package")
 
     def test_bytecode_package(self):
         p = self._make_package("my_package")